requirements.txt
cryptography==42.0.5
Django==5.0.4
django-extensions==3.2.3
djangorestframework==3.15.1
ipdb==0.13.13
ipython==8.23.0
pika==1.3.2
pyjwt==2.8.0
pytest==8.1.1
pytest-django==4.8.0
python-dotenv==1.0.1
